requirements.txt
flask==1.1.1
sqlalchemy==1.3.10
flask-sqlalchemy==2.4.1
flask-restplus==0.13.0
Flask-Cors==3.0.8
python-dotenv==0.10.3
pymysql==0.9.3
flask-migrate==2.5.2
marshmallow==3.2.1
flask_accepts==0.14.2
flask-script==2.0.6
requests==2.22.0
retry==0.9.2
werkzeug==0.16.1
six==1.14.0
gunicorn==20.0.4
Flask-Bcrypt==0.7.1
pyjwt==1.7.1
